fix: 보너스 투구 출력부분 수정
