+++ src/main/java/bowling/domain/status/Spare.java
@@ -3,6 +3,8 @@ package bowling.domain.status;
 import bowling.domain.Pitch;
 
 public class Spare extends Finished {
+    private static final int BONUS_PITCH_COUNT = 1;
+
     private final Pitch current;
 
     public Spare(Pitch current) {
@@ -13,7 +15,11 @@ public class Spare extends Finished {
     public Status roll(int fallenPins) {
         Pitch pitch = new Pitch(fallenPins);
 
-        return new Final(pitch, 0);
+        if (pitch.isStrike()) {
+            return new Bonus(pitch, BONUS_PITCH_COUNT, new Strike(pitch));
+        }
+
+        return new Bonus(pitch, BONUS_PITCH_COUNT, new Hold(pitch));
     }
 
     @Override
